Bug Fixes and Multiple Keypoints (#10)
## Summary of Changes
In this change, I have modified the Article and ChartSection components
to render multiple key points on each rendered chart.

## Motivation
As we scroll through the article, we still want to be able to see key
points that are not in focus. This change will highlight points that are
not in focus with a yellow circle and points that are in focus with a
red circle.