Add Article Customizations (#11)
## Summary of Changes
In this PR, I have included a few changes to introduce article
customizations for authors. In particular, this change includes
- the ability to change the title, byline, and date
- the ability to change the color of text, background, line color, key
point colors
- re-organized code to include a new `ArticleContainer` component that
contains an `Article` and a `Builder`
- types stored in an `ArticleTypes` file
- constants stored in an `ArticleConstants` file

## Motivation
We want authors to be able to customize their articles to be in line
with their publishing platforms. These changes increase the amount of
flexibility authors have and make this product more professional.