Add Section Builder (#14)
## Summary of Changes
In this PR, I have included changes to build article sections with the
key points that we collect from the user. This includes new
`SectionBuilder` and modified `Chart` and `TextSection` components.

## Motivation
By allowing users to create sections, we can get the generated article
to look more like a traditional article. This is also a good
intermediary step before adding `linking` and `scrollytelling` features
to the generated article.