Support Bivariate Data (#37)
## Summary of Changes
In this pull request, I have included changes to the Article components
to support bivariate data.

## Motivation
We want to be able to visualize and generate articles for complex
datasets with our article authoring tool. By supporting bivariate data,
we are able to visualize more than one trend in a dataset. This also
paves the way for supporting n-variate data.